+++ build.gradle
@@ -11,7 +11,7 @@ java {
 }
 
 group = 'cloud.eppo'
-version = '3.1.1-SNAPSHOT'
+version = '4.0.0-SNAPSHOT'
 ext.isReleaseVersion = !version.endsWith("SNAPSHOT")
 
 import org.apache.tools.ant.filters.ReplaceTokens
@@ -30,8 +30,7 @@ repositories {
 }
 
 dependencies {
-  // Re-export classes and interfaces that will be used upstream
-  api 'cloud.eppo:sdk-common-jvm:3.3.1'
+  api 'cloud.eppo:sdk-common-jvm:3.5.0'
 
   implementation 'com.github.zafarkhaja:java-semver:0.10.2'
   implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.1'
@@ -40,12 +39,12 @@ dependencies {
   // Logback classic 1.3.x is compatible with java 8
   implementation 'ch.qos.logback:logback-classic:1.3.14'
 
-  testImplementation 'cloud.eppo:sdk-common-jvm:3.2.0-SNAPSHOT:tests'
+  testImplementation 'cloud.eppo:sdk-common-jvm:3.5.0-SNAPSHOT:tests'
   testImplementation platform('org.junit:junit-bom:5.10.2')
   testImplementation 'org.junit.jupiter:junit-jupiter'
   testImplementation 'com.github.tomakehurst:wiremock-jre8:2.35.2'
   testImplementation 'org.mockito:mockito-core:4.11.0'
-  testImplementation 'com.squareup.okhttp3:okhttp:4.9.1'
+  testImplementation 'com.squareup.okhttp3:okhttp:4.12.0'
 }
 
 test {
