+++ src/main/java/cloud/eppo/AppDetails.java
@@ -1,4 +1,4 @@
-package com.eppo.sdk.helpers;
+package cloud.eppo;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -6,20 +6,20 @@ import java.util.Properties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class AppDetails {
+class AppDetails {
   private static final Logger log = LoggerFactory.getLogger(AppDetails.class);
   static AppDetails instance;
   private final String version;
   private final String name;
 
-  public static AppDetails getInstance() {
+  static AppDetails getInstance() {
     if (instance == null) {
       instance = new AppDetails();
     }
     return instance;
   }
 
-  public AppDetails() {
+  AppDetails() {
     Properties prop = new Properties();
     try {
       prop = readPropertiesFile("app.properties");
@@ -30,15 +30,15 @@ public class AppDetails {
     this.name = prop.getProperty("app.name", "java-server-sdk");
   }
 
-  public String getVersion() {
+  String getVersion() {
     return this.version;
   }
 
-  public String getName() {
+  String getName() {
     return this.name;
   }
 
-  public static Properties readPropertiesFile(String fileName) throws IOException {
+  static Properties readPropertiesFile(String fileName) throws IOException {
     ClassLoader loader = Thread.currentThread().getContextClassLoader();
     Properties props = new Properties();
     InputStream resourceStream = loader.getResourceAsStream(fileName);
